requirements.txt
streamlit
lancedb
pandas
sentence-transformers
transformers
torch
pillow
